requirements.txt
# LLM Consultant Bot - Основные зависимости
# Telegram Bot API
aiogram>=3.4.0,<4.0.0

# OpenAI/OpenRouter API клиент
openai>=1.12.0,<2.0.0

# Конфигурация и окружение
pyyaml>=6.0.1,<7.0.0
python-dotenv>=1.0.0,<2.0.0

# Разработка и тестирование (опционально)
pytest>=7.4.0,<8.0.0

# Дополнительные зависимости для стабильной работы
# Асинхронные HTTP запросы (используется aiogram)
aiohttp>=3.9.0,<4.0.0

# Сертификаты для HTTPS запросов
certifi>=2023.0.0

# Поддержка асинхронности
asyncio-mqtt>=0.16.0  # если понадобится MQTT в будущем